CI : Add image validation
This compares graphics and documentation images with a "known good" set of images, producing an HTML report uploaded as an artifact named "imageCheck". At this time, that "known good" set are images extracted from the latest Gaffer release for the major version matching the current push, pull request, or release (based on the `GITHUB_REF_NAME` environment variable provided by CI).

This approach may prove to accumulate false-positive tests while changes are in a merged but unreleased state, but as this already goes a long way to providing much needed visibility into accidental changes, it's worth living with to see how badly we are affected by that churn.