Docs : Adjust bookmark annotations
Make outline taller to account for differences in fonts and layout of menu items between the gcc9 and gcc11 desktop environments. This results in a slightly unhappy meduium where the single outline works for both but is not great for either, but we'll be able to tidy this once we drop the gcc9 builds.

We also offset the outlines on Windows to account for additional "Drives" bookmark item.