Menu : Work around slow QAction connections in Qt 6
Connecting to QAction.toggled and QAction.triggered has become more expensive with Qt 6. This is very noticeable when repeatedly searching in the NodeMenu where each keypress could be creating hundreds of new QActions as the search menu is updated. This overhead appears to increase as more QActions are created over time, so repeatedly opening the NodeMenu and pressing "c" to search for nodes containing "c" gets progressively slower.

To avoid creating these connections in the first place, this replaces the QAction.triggered connections with a single connection to QMenu.triggered, which is emitted whenever an action in the menu is triggered. This still leaves the individual QAction.toggled connections, but those are few and far between, and are not required for NodeMenu entries. It would be a bit cleaner if we could also remove the QAction.toggled connections as it seems QAction.toggled is emitted alongside QMenu.triggered when the QAction is triggered...