RenderPassEditor : Add column for directly adding favourites
This provides an alternative to right-clicking on a column header to add it as a favourite, and allows adding options that haven't already been registered as columns. Which can be quite useful, as the current column registrations aren't exhaustive.

We want the "+" icon of the column header to feel more like a button, but this seems a bit tricky in Qt. Currently we're leveraging `sectionPressed` to respond to left-clicks within the header, but this doesn't prevent the section from being sorted. We'd also like the icon to change on mouse hover. A later commit will provide some improvements via a custom QHeaderView.